Shot Animation implemented
New method Tank::shot() which sets the flag and starts the clock responsible for drawing shot on the map. If time elapsed is greater than threshold the flag is set to false.

Tank requires new texture in the constructor, which is responsible for
shot animation.

It is possible to configure shot animation by 2 variables defined in
tank.hpp:
- shot animation distance (how far the animation is from the barrel)
- shot animation duration (how long will the animation be displayed)

Code restructure to honour Single Responsibility Principle (draw function
does not update objects any more). Because of this change it was
necessary to update TankTests aswell. Dummy RenderWindow is no longer
necessary in tests.