Dynamic tank acceleration implemented
New module 'Engine' responsible for tank movement implemented.

Currently there is one concrete implementation called 'SquareRootEngine',
acceleration in this engine is based on 'square root' function. This
engine has one parameter which controls how fast the tank accelerates
(stepCount - this is the number of update frames that have to be called
until tank reaches maximum speed).

Tank can be created with many types of engines.

For debug purposes accelerometer was added that displays current tank
speed on the screen.