Trace class added
Trace can be used instead of sf::Sprite. It is more lightweight, because Trace does not need to be transformed once it is positioned, the only things that are necessary is: increase / decrease the trace length. Thanks to using sf::VertexArray it is possible to easily decrease the trace length, which can be used to decay the traces